Refactor to maven project
