+++ .gitignore
@@ -5,3 +5,9 @@
 *.war
 *.ear
 /.project
+
+.DS_Store
+/target/
+.classpath
+.settings/
+bin/
