+++ src/main/java/com/hasdaipacheco/screentask/Resources.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package screentask;
+package com.hasdaipacheco.screentask;
 
 import java.io.BufferedReader;
 import java.io.File;
