+++ src/main/resources/WebServer/cursor_arrow.png
